requirements.txt
idna==3.10
python-telegram-bot==21.11.1
requests==2.32.3
sniffio==1.3.1
urllib3==2.3.0
